+++ packages/markdown/src/index.ts
@@ -14,6 +14,8 @@ export * from './document/index.js';
 export * from './plugin/index.js';
 export * from './parser/index.js';
 export * from './serializer/index.js';
+export * from './markdown/index.js';
+export * from './markdown/index.js';
 export * from './render/index.js';
 export * from './stream/index.js';
 export * from './plugins/index.js';